+++ tests/creature.rs
@@ -77,6 +77,8 @@ fn statistics_generated() {
         type Statistic = EmptyStat;
         type StatisticsSeed = u32;
         type StatisticsAlteration = ();
+        type Status = EmptyStatus;
+        type StatusesAlteration = ();
 
         fn generate_statistics(
             &self,
@@ -119,6 +121,8 @@ fn regenerate_statistics() {
         // Vec with pair (id, value).
         type StatisticsSeed = Vec<(u32, u32)>;
         type StatisticsAlteration = ();
+        type Status = EmptyStatus;
+        type StatusesAlteration = ();
 
         fn generate_statistics(
             &self,
@@ -347,6 +351,8 @@ fn user_metrics() {
         type Statistic = SimpleStatistic<u32, u64>;
         type StatisticsSeed = u64;
         type StatisticsAlteration = ();
+        type Status = EmptyStatus;
+        type StatusesAlteration = ();
 
         fn generate_statistics(
             &self,
@@ -532,6 +538,8 @@ fn remove_creature_on_alter() {
         type Statistic = EmptyStat;
         type StatisticsSeed = ();
         type StatisticsAlteration = ();
+        type Status = EmptyStatus;
+        type StatusesAlteration = ();
 
         fn alter_statistics(
             &self,