+++ CHANGELOG.md
@@ -7,17 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for status effects.
-- New method `generate_status` in `CharacterRules`.
+- New methods `generate_status` and `alter_statuses` in `CharacterRules`.
 - New methods `apply_status`, `update_status` and `delete_status` in `FightRules`.
 - `InflictStatus` and `ClearStatus` events.
 - Added `StatusNotPresent` to `WeaselError`.
 - Mutable iterators over statistics and abilities.
 - New event `EnvironmentRound`.
+- New associated type `Potency` in `FightRules`. 
+- New associated types `Status` and `StatusesAlteration` in `CharacterRules`.
 - Example to showcase status effects.
 
 ### Changed
 - Renamed `ActorRules`'s `alter` into `alter_abilities` and `CharacterRules`'s `alter` into `alter_statistics`.
-- New associated types `Status` and `Potency` in `FightRules`. 
 
 ## [0.5.0] - 2020-02-26
 ### Added
@@ -40,10 +41,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduced inanimate objects.
 - New events `CreateObject` and `RemoveObject`.
 - Improved public API for `Battle` and its submodules.
+- New associated type `ObjectId` in `CharacterRules`.
 
 ### Changed
 - It's now possible to manually set an event's origin.
-- New associated type `ObjectId` in `CharacterRules`.
 
 ## [0.3.1] - 2020-02-17
 ### Added